requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
rich>=13.5.0
textstat>=0.7.3
python-Levenshtein>=0.23.0
numpy>=1.25.0
matplotlib>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
scipy>=1.10.0
scikit-learn>=1.0.0
pandas>=1.0.0
seaborn>=0.11.0 